Add download stats button on profile page
