--- src/components/DownloadStatsPage.tsx
@@ -1,1 +1,0 @@
-
